+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
-## 1.0.0 (Unreleased)
+## 2.0.0
 
-Initial release.
+- Update for Siemens Industrial Edge App IIH Essentials
+- Add multiple tags query
+- Add where clause to query
+- New query editor
+- Add editor for Grafana Dashboard Variables
+
+## 1.0.0
+
+- Version for Siemens Industrial Edge App Data Service